chore: setup svgr (#18)
* install dependencies

* update next.config.mjs

* create svgr.d.ts

* update tsconfig.json

* update package.json

removed usage of turbopack to improve compatibility with serwist
(which does not yet support turbopack). also, as of writing, the
installation guide for svgr is configured with webpack

see:
* https://github.com/serwist/serwist/issues/54
* https://github.com/gregberge/svgr/blob/180eb6d/website/pages/docs/next.mdx#usage

* update next.config.mjs

opted to use the working implementation in the svgr documentation rather
than the related issue as it lead to some build issue. also refactored
the config file for clarity.